+++ backend/server.py
@@ -7,5 +7,5 @@ app = create_app()
 if __name__ == '__main__':
     app.run(
         host="0.0.0.0",
-        debug=(os.getenv('APP_ENVIRONMENT') == 'local')
+        debug=(os.getenv('APP_ENVIRONMENT') == 'debug')
     )