+++ frontend/src/components/Results/Criterias/CriteriaRedirects.vue
@@ -2,31 +2,23 @@
     <BasicAccordion :status="result.accepted">
         <template #title>
             <p class="text-gray-900 dark:text-gray-200">
-                Criteria {{ result.id }}: Outgoing requests
+                Criteria {{ result.id }}: Redirects
             </p>
         </template>
         <template #details>
-            <div class="overflow-x-auto">
+            <div v-if="result.details.redirects.length === 0">
+                Looks good! There are no redirects.
+            </div>
+            <div v-else class="overflow-x-auto">
                 <table class="w-full border-separate border-spacing-1">
                     <thead class="dark:bg-opacity-25 bg-gray-800">
                         <tr>
-                            <th class="border border-gray-600 px-2">Response</th>
-                            <th class="border border-gray-600 px-2">Method</th>
                             <th class="border border-gray-600 px-2">URL</th>
                             <th class="border border-gray-600 px-2">Date</th>
                             <th class="border border-gray-600 px-2">Content-Type</th>
-                            <th class="border border-gray-600 px-2">Size</th>
                         </tr>
                     </thead>
-                    <tr v-for="request in result.details.requests" class="text-gray-800 dark:text-gray-200">
-                        <td class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25">
-                            <CheckCircleIcon v-if="requestSuccess(request)"
-                                class="w-8 md:w-10 h-auto m-auto text-emerald-600" />
-                            <XCircleIcon v-if="!requestSuccess(request)" class="w-10 h-auto m-auto text-rose-600" />
-                        </td>
-                        <td class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25">
-                            {{ request.method }}
-                        </td>
+                    <tr v-for="request in result.details.redirects" class="text-gray-800 dark:text-gray-200">
                         <td
                             class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25 text-left pl-4 max-w-sm">
                             <a :href="request.url" class="hover:underline break-all ">
@@ -44,17 +36,6 @@
                                 No Response
                             </div>
                         </td>
-                        <td class="border border-gray-700 dark:border-gray-100 dark:border-opacity-25">
-                            <div v-if="!hasResponse(request)">
-                                No Response
-                            </div>
-                            <div v-else-if="isRedirect(request)">
-                                Redirect
-                            </div>
-                            <div v-else>
-                                {{ convert(request.size) }}
-                            </div>
-                        </td>
                     </tr>
                 </table>
             </div>
@@ -63,12 +44,9 @@
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
-import type { PropType } from 'vue'
+import { defineComponent, PropType } from 'vue'
 import { CriteriaModel, RequestModel } from '../../../models/result';
-import convertBytesToString from '../../../utils/functions'
-
-import { CheckCircleIcon, XCircleIcon } from '@heroicons/vue/24/outline'
+import convertBytesToString from '../../../utils/functions';
 
 export default defineComponent({
     props: {
@@ -94,15 +72,6 @@ export default defineComponent({
         hasResponse(request: RequestModel) {
             return request.response !== null
         },
-        isRedirect(request: RequestModel) {
-            if (request.response) {
-                return (request.response.status_code === 301)
-            }
-            return false
-        }
     },
-    components: {
-        CheckCircleIcon, XCircleIcon
-    }
 })
 </script>